refactor(#98): 공통 함수 utils로 분리
date를 string으로 변환하는 함수와 datetime을 string으로 변환하는 함수는 추후 사용될 수도 있다고 판단하여 utils의 converters.ts 파일로 옮겨두었습니다.
