+++ client/src/utils/converters.ts
@@ -16,3 +16,18 @@ export const stringTuplesToDateRanges = (dates: string[][]) => {
     .filter(([start, end]) => start <= end) as DateRange[];
 };
 
+export const dateToString = (date: Date) => {
+  const year = date.getFullYear();
+  const month = (date.getMonth() + 1).toString().padStart(2, '0');
+  const day = date.getDate().toString().padStart(2, '0');
+  return `${year}-${month}-${day}`;
+};
+
+
+// DateTime을 string으로 변환
+export const dateTimeToString = (dateTime: Date) => {
+  const date = dateToString(dateTime);
+  const hours = dateTime.getHours().toString().padStart(2, '0');
+  const time = `${hours}:00:00`
+  return `${date}T${time}`;
+};
